requirements.txt
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
dnspython==2.2.1
exceptiongroup==1.0.4
Flask==2.2.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
pydantic==1.10.2
pymongo==4.3.3
pytest==7.2.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
uuid==1.30
Werkzeug==2.2.2
zipp==3.11.0
